requirements.txt
Django==1.11.2
djangorestframework==3.6.3
gunicorn==19.7.1
pytz==2017.2
